% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{to}
\alias{date}
\alias{date.gmail_draft}
\alias{date.gmail_message}
\alias{from}
\alias{from.gmail_draft}
\alias{from.gmail_message}
\alias{to}
\alias{to.gmail_draft}
\alias{to.gmail_message}
\title{Methods to get values from message or grafts or set values in a mime message}
\usage{
to(x, ...)

\method{to}{gmail_message}(x, ...)

\method{to}{gmail_draft}(x, ...)

from(x, ...)

\method{from}{gmail_message}(x, ...)

\method{from}{gmail_draft}(x, ...)

date(x, ...)

\method{date}{gmail_message}(x, ...)

\method{date}{gmail_draft}(x, ...)
}
\arguments{
\item{x}{the object from which to get or set the field}

\item{...}{other parameters passed to methods}

\item{val}{value to set}

\item{vals}{values to set, they will be joined by commas}
}
\description{
Methods to get values from message or grafts or set values in a mime message
}
\examples{
\dontrun{
to(my_message)
to(my_draft)
}
}

