\name{anRpackage-package}
\alias{anRpackage-package}
\alias{anRpackage}
\docType{package}
\title{
Send gmail with attachments
}
\description{
Send gmail with attachments
}
\details{
\tabular{ll}{
Package: \tab qdap\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-01\cr
License: \tab GPL-2\cr
}
}
\author{
Tyler Rinker

Maintainer: Tyler Rinker <tyler.rinker@gmail.com>
}
