% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{as.character.mime_message}
\alias{as.character.mime_message}
\alias{as.character.mime_part}
\title{Convert a mime object to character representation}
\usage{
\method{as.character}{mime_message}(x, ...)

\method{as.character}{mime_part}(x, ...)
}
\arguments{
\item{x}{object to convert}

\item{...}{additional arguments ignored}
}
\description{
These functions convert a mime object into the final mime character
representation.
}

