% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{body.gmail_message}
\alias{body.gmail_message}
\title{Extract the message body from an email message}
\usage{
\method{body}{gmail_message}(x, collapse = FALSE, ...)
}
\arguments{
\item{x}{message to retrieve body for}

\item{collapse}{collapse multipart message into one}

\item{...}{other options ignored}
}
\description{
If a multipart message was returned each part will be a separate list item.
}

