% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{body}
\alias{body}
\alias{body.gmail_draft}
\alias{body.gmail_message}
\title{Get the body text of a message or draft}
\usage{
body(x, ...)

\method{body}{gmail_message}(x, collapse = FALSE, ...)

\method{body}{gmail_draft}(x, ...)
}
\arguments{
\item{x}{the object from which to retrieve the body}

\item{...}{other parameters passed to methods}

\item{collapse}{collapse multipart message into one}

\item{x}{message to retrieve body for}

\item{...}{other options ignored}
}
\description{
Get the body text of a message or draft

If a multipart message was returned each part will be a separate list item.
}
\examples{
\dontrun{
body(my_message)
body(my_draft)
}
}

