\name{cell.ext}
\alias{cell.ext}
\docType{data}
\title{
A listing of cell phone extensions
}
\description{
Adding these extensions to a cell phone number allows the user to send emails to cell phones as a text.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
http://www.emailtextmessages.com/
}
