% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{to.mime_message}
\alias{bcc}
\alias{cc}
\alias{common_fields}
\alias{from.mime_message}
\alias{subject.mime_message}
\alias{to.mime_message}
\title{Accessor functions for mime messages}
\usage{
\method{to}{mime_message}(x, vals, ...)

\method{from}{mime_message}(x, val, ...)

cc(x, vals, ...)

bcc(x, vals, ...)

\method{subject}{mime_message}(x, val, ...)
}
\arguments{
\item{x}{the object whose fields you are setting}

\item{vals}{one or more values to use, will be joined by commas}

\item{...}{}

\item{val}{the value to set}
}
\description{
Accessor functions for mime messages
}

