% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/draft.R
\name{create_draft}
\alias{create_draft}
\title{Create a draft from a mime message}
\usage{
create_draft(mail, user_id = "me", type = c("multipart", "media",
  "resumable"))
}
\arguments{
\item{mail}{mime mail message created by mime}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}

\item{type}{the type of upload to perform}
}
\description{
Create a draft from a mime message
}
\examples{
\dontrun{
create_draft(mime(from="you@me.com", to="any@one.com",
                          subject='hello", "how are you doing?"))
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/drafts/create}
}

