% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{create_label}
\alias{create_label}
\title{Create a new label}
\usage{
create_label(name, label_list_visibility = c("hide", "show", "show_unread"),
  message_list_visibility = c("hide", "show"), user_id = "me")
}
\arguments{
\item{name}{name to give to the new label}

\item{label_list_visibility}{The visibility of the label in the label list in the Gmail web interface.}

\item{message_list_visibility}{The visibility of messages with this label in the message list in the Gmail web interface.}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to create a label.
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/labels/create}
}
\seealso{
Other label: \code{\link{delete_label}},
  \code{\link{labels}}, \code{\link{label}},
  \code{\link{update_label}}
}

