% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{date}
\alias{date}
\alias{date.gmail_draft}
\alias{date.gmail_message}
\title{Get the date field of a message or draft}
\usage{
date(x, ...)

\method{date}{gmail_message}(x, ...)

\method{date}{gmail_draft}(x, ...)
}
\arguments{
\item{x}{the object from which to retrieve the field}

\item{...}{other parameters passed to methods}
}
\description{
Get the date field of a message or draft
}
\examples{
\dontrun{
date(my_message)
date(my_draft)
}
}

