% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{delete_message}
\alias{delete_message}
\title{Permanently delete a single message}
\usage{
delete_message(id, user_id = "me")
}
\arguments{
\item{id}{message id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to delete a given message by id.  This cannot be undone!
}
\examples{
\dontrun{
delete_message('12345')
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/messages/delete}
}
\seealso{
Other message: \code{\link{attachment}},
  \code{\link{import_message}},
  \code{\link{insert_message}}, \code{\link{messages}},
  \code{\link{message}}, \code{\link{modify_message}},
  \code{\link{save_attachments}},
  \code{\link{save_attachment}},
  \code{\link{send_message}}, \code{\link{trash_message}},
  \code{\link{untrash_message}}
}

