% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{delete_thread}
\alias{delete_thread}
\title{Permanently delete a single Thread.}
\usage{
delete_thread(id, user_id = "me")
}
\arguments{
\item{id}{thread id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to delete a given Thread by id.  This cannot be undone!
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/threads/delete}
}

