% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/gmailr.R
\name{format}
\alias{format}
\alias{format.gmail_draft}
\alias{format.gmail_drafts}
\alias{format.gmail_message}
\alias{format.gmail_messages}
\alias{format.gmail_thread}
\alias{format.gmail_threads}
\title{Format gmailr objects for pretty printing}
\usage{
\method{format}{gmail_message}(x, ...)

\method{format}{gmail_thread}(x, ...)

\method{format}{gmail_draft}(x, ...)

\method{format}{gmail_messages}(x, ...)

\method{format}{gmail_threads}(x, ...)

\method{format}{gmail_drafts}(x, ...)
}
\arguments{
\item{x}{object to format}

\item{...}{additional arguments ignored}
}
\description{
Format gmailr objects for pretty printing
}

