% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{from}
\alias{from}
\title{Get the from field of a message or draft}
\usage{
from(x, ...)
}
\arguments{
\item{x}{the object from which to retrieve the field}

\item{...}{other parameters passed to methods}
}
\description{
Get the from field of a message or draft
}
\examples{
\dontrun{
from(my_message)
from(my_draft)
}
}

