% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_message.R
\name{gm_attachments}
\alias{gm_attachments}
\title{Retrieve information about attachments}
\usage{
gm_attachments(x, ...)
}
\arguments{
\item{x}{An object from which to retrieve the attachment information.}

\item{...}{other parameters passed to methods}
}
\value{
A data.frame with the \code{filename}, \code{type}, \code{size} and \code{id} of each
attachment in the message.
}
\description{
Retrieve information about attachments
}
