% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm-auth.R
\name{gm_auth_configure}
\alias{gm_auth_configure}
\alias{gm_auth_config}
\alias{gm_oauth_app}
\title{Edit auth configuration}
\usage{
gm_auth_configure(app, path)

gm_oauth_app()
}
\arguments{
\item{app}{OAuth app.}

\item{path}{JSON downloaded from Google Cloud Platform Console, containing a
client id (aka key) and secret, in one of the forms supported for the \code{txt}
argument of \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} (typically, a file path or JSON string).}
}
\value{
* `gm_auth_configure()`: An object of R6 class [gargle::AuthState],
    invisibly.
  * `gm_oauth_app()`: the current user-configured [httr::oauth_app()].
}
\description{
These functions give the user more control over auth than what is possible
with [gm_auth()]. `gm_auth_configure()` gives control of:
  * The OAuth app, which is used when obtaining a user token.

See the vignette [How to get your own API
credentials](https://gargle.r-lib.org/articles/get-api-credentials.html) for
more.
}
\examples{
# see the current user-configured OAuth app (probaby `NULL`)
gm_oauth_app()

if (require(httr)) {

  # store current state, so we can restore
  original_app <- gm_oauth_app()

  # bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "123456789.apps.googleusercontent.com",
    secret = "abcdefghijklmnopqrstuvwxyz"
  )
  gm_auth_configure(app = google_app)

  # confirm current app
  gm_oauth_app()

  # restore original state
  gm_auth_configure(app = original_app)
  gm_oauth_app()
}

\dontrun{
# bring your own app via JSON downloaded from Google Developers Console
gm_auth_configure(
  path = "/path/to/the/JSON/you/downloaded/from/google/dev/console.json"
)
}

}
\seealso{
Other auth functions: \code{\link{gm_auth}},
  \code{\link{gm_deauth}}
}
\concept{auth functions}
