% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{gm_create_draft}
\alias{gm_create_draft}
\title{Create a draft from a mime message}
\usage{
gm_create_draft(mail, user_id = "me")
}
\arguments{
\item{mail}{mime mail message created by mime}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Create a draft from a mime message
}
\examples{
\dontrun{
gm_create_draft(gm_mime(From="you@me.com", To="any@one.com",
                          Subject="hello", "how are you doing?"))
}
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.drafts/create}
}
