% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{gm_create_label}
\alias{gm_create_label}
\title{Create a new label}
\usage{
gm_create_label(
  name,
  label_list_visibility = c("show", "hide", "show_unread"),
  message_list_visibility = c("show", "hide"),
  user_id = "me"
)
}
\arguments{
\item{name}{name to give to the new label}

\item{label_list_visibility}{The visibility of the label in the label list in the Gmail web interface.}

\item{message_list_visibility}{The visibility of messages with this label in the message list in the Gmail web interface.}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to create a label.
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.labels/create}
}
\seealso{
Other label: 
\code{\link{gm_delete_label}()},
\code{\link{gm_labels}()},
\code{\link{gm_label}()},
\code{\link{gm_update_label}()}
}
\concept{label}
