% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm-auth.R
\name{gm_deauth}
\alias{gm_deauth}
\title{Clear current token}
\usage{
gm_deauth()
}
\description{
Clears any currently stored token. The next time gmailr needs a token,
the token acquisition process starts over, with a fresh call to
\code{\link[=gm_auth]{gm_auth()}} and, therefore, internally, a call to
\code{\link[gargle:token_fetch]{gargle::token_fetch()}}. Unlike some other packages that use gargle,
gmailr is not usable in a de-authorized state. Therefore, calling
\code{gm_deauth()} only clears the token, i.e. it does NOT imply that
subsequent requests are made with an API key in lieu of a token.
}
\examples{
\dontrun{
gm_deauth()
}
}
\seealso{
Other auth functions: 
\code{\link{gm_auth_configure}()},
\code{\link{gm_scopes}()}
}
\concept{auth functions}
