% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{gm_delete_draft}
\alias{gm_delete_draft}
\title{Permanently delete a single draft}
\usage{
gm_delete_draft(id, user_id = "me")
}
\arguments{
\item{id}{message id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to delete a given draft by id.  This cannot be undone!
}
\examples{
\dontrun{
delete_draft('12345')
}
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.drafts/delete}
}
\seealso{
Other draft: 
\code{\link{gm_drafts}()},
\code{\link{gm_draft}()},
\code{\link{gm_send_draft}()}
}
\concept{draft}
