% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{gm_delete_label}
\alias{gm_delete_label}
\title{Permanently delete a label}
\usage{
gm_delete_label(id, user_id = "me")
}
\arguments{
\item{id}{label id to retrieve}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to delete a label by id.  This cannot be undone!
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.labels/delete}
}
\seealso{
Other label: 
\code{\link{gm_create_label}()},
\code{\link{gm_labels}()},
\code{\link{gm_label}()},
\code{\link{gm_update_label}()}
}
\concept{label}
