% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_thread.R
\name{gm_delete_thread}
\alias{gm_delete_thread}
\title{Permanently delete a single thread.}
\usage{
gm_delete_thread(id, user_id = "me")
}
\arguments{
\item{id}{thread id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to delete a given thread by id.  This cannot be undone!
}
\examples{
\dontrun{
gm_delete_thread(12345)
}
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.threads/delete}
}
\seealso{
Other thread: 
\code{\link{gm_modify_thread}()},
\code{\link{gm_threads}()},
\code{\link{gm_thread}()},
\code{\link{gm_trash_thread}()},
\code{\link{gm_untrash_thread}()}
}
\concept{thread}
