% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{gm_draft}
\alias{gm_draft}
\title{Get a single draft}
\usage{
gm_draft(id, user_id = "me", format = c("full", "minimal", "raw"))
}
\arguments{
\item{id}{draft id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}

\item{format}{format of the draft returned}
}
\description{
Function to retrieve a given draft by <-
}
\examples{
\dontrun{
my_draft = gm_draft('12345')
}
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.drafts/get}
}
\seealso{
Other draft: 
\code{\link{gm_delete_draft}()},
\code{\link{gm_drafts}()},
\code{\link{gm_send_draft}()}
}
\concept{draft}
