% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_draft.R
\name{gm_drafts}
\alias{gm_drafts}
\title{Get a list of drafts}
\usage{
gm_drafts(num_results = NULL, page_token = NULL, user_id = "me")
}
\arguments{
\item{num_results}{the number of results to return.}

\item{page_token}{retrieve a specific page of results}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Get a list of drafts possibly matching a given query string.
}
\examples{
\dontrun{
my_drafts <- gm_drafts()

first_10_drafts <- gm_drafts(10)
}
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.drafts/list}
}
\seealso{
Other draft: 
\code{\link{gm_delete_draft}()},
\code{\link{gm_draft}()},
\code{\link{gm_send_draft}()}
}
\concept{draft}
