% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_auth.R
\name{gm_has_token}
\alias{gm_has_token}
\title{Is there a token on hand?}
\usage{
gm_has_token()
}
\value{
Logical.
}
\description{
Reports whether gmailr has stored a token, ready for use in downstream
requests.
}
\examples{
gm_has_token()
}
\seealso{
Other low-level API functions: 
\code{\link{gm_token}()}
}
\concept{low-level API functions}
