% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{gm_label}
\alias{gm_label}
\title{Get a specific label}
\usage{
gm_label(id, user_id = "me")
}
\arguments{
\item{id}{label id to retrieve}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Get a specific label by id and user_id.
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.labels/get}
}
\seealso{
Other label: 
\code{\link{gm_create_label}()},
\code{\link{gm_delete_label}()},
\code{\link{gm_labels}()},
\code{\link{gm_update_label}()}
}
\concept{label}
