% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_message.R
\name{gm_message}
\alias{gm_message}
\title{Get a single message}
\usage{
gm_message(
  id,
  user_id = "me",
  format = c("full", "metadata", "minimal", "raw")
)
}
\arguments{
\item{id}{message id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}

\item{format}{format of the message returned}
}
\description{
Function to retrieve a given message by id
}
\examples{
\dontrun{
my_message <- gm_message(12345)
}
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.messages}
}
\seealso{
Other message: 
\code{\link{gm_attachment}()},
\code{\link{gm_delete_message}()},
\code{\link{gm_import_message}()},
\code{\link{gm_insert_message}()},
\code{\link{gm_messages}()},
\code{\link{gm_modify_message}()},
\code{\link{gm_save_attachments}()},
\code{\link{gm_save_attachment}()},
\code{\link{gm_send_message}()},
\code{\link{gm_trash_message}()},
\code{\link{gm_untrash_message}()}
}
\concept{message}
