% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_auth.R
\name{gm_oauth_app}
\alias{gm_oauth_app}
\title{Get currently configured OAuth app (deprecated)}
\usage{
gm_oauth_app()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

In light of the new \code{\link[gargle:gargle_oauth_client_from_json]{gargle::gargle_oauth_client()}} constructor and class of
the same name, \code{gm_oauth_app()} is being replaced by \code{\link[=gm_oauth_client]{gm_oauth_client()}}.
}
\keyword{internal}
