% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{gm_send_draft}
\alias{gm_send_draft}
\title{Send a draft}
\usage{
gm_send_draft(draft, user_id = "me")
}
\arguments{
\item{draft}{the draft to send}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Send a draft to the recipients in the To, CC, and Bcc headers.
}
\examples{
\dontrun{
draft <- gm_create_draft(gm_mime(From="you@me.com", To="any@one.com",
                      Subject="hello", "how are you doing?"))
gm_send_draft(draft)
}
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.drafts/send}
}
\seealso{
Other draft: 
\code{\link{gm_delete_draft}()},
\code{\link{gm_drafts}()},
\code{\link{gm_draft}()}
}
\concept{draft}
