% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{gm_thread}
\alias{gm_thread}
\title{Get a single thread}
\usage{
gm_thread(id, user_id = "me")
}
\arguments{
\item{id}{thread id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to retrieve a given thread by id
}
\examples{
\dontrun{
my_thread = gm_thread(12345)
}
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.threads}
}
\seealso{
Other thread: 
\code{\link{gm_delete_thread}()},
\code{\link{gm_modify_thread}()},
\code{\link{gm_threads}()},
\code{\link{gm_trash_thread}()},
\code{\link{gm_untrash_thread}()}
}
\concept{thread}
