% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm-auth.R
\name{gm_token}
\alias{gm_token}
\title{Produce configured token}
\usage{
gm_token()
}
\value{
A `request` object (an S3 class provided by [httr][httr::httr]).
}
\description{
For internal use or for those programming around the Gmail API.
Returns a token pre-processed with [httr::config()]. Most users
do not need to handle tokens "by hand" or, even if they need some
control, [gm_auth()] is what they need. If there is no current
token, [gm_auth()] is called to either load from cache or
initiate OAuth2.0 flow.
If auth has been deactivated via [gm_deauth()], `gm_token()`
returns `NULL`.
}
\examples{
\dontrun{
gm_token()
}
}
\concept{low-level API functions}
