% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{gm_trash_thread}
\alias{gm_trash_thread}
\title{Send a single thread to the trash}
\usage{
gm_trash_thread(id, user_id = "me")
}
\arguments{
\item{id}{thread id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to trash a given thread by id.  This can be undone by \code{\link[=gm_untrash_thread]{gm_untrash_thread()}}.
}
\examples{
\dontrun{
trash_thread(12345)
}
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.threads/trash}
}
\seealso{
Other thread: 
\code{\link{gm_delete_thread}()},
\code{\link{gm_modify_thread}()},
\code{\link{gm_threads}()},
\code{\link{gm_thread}()},
\code{\link{gm_untrash_thread}()}
}
\concept{thread}
