% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_label.R
\name{gm_update_label}
\alias{gm_update_label}
\alias{gm_update_label_patch}
\title{Update a existing label.}
\usage{
gm_update_label(id, label, user_id = "me")

gm_update_label_patch(id, label, user_id = "me")
}
\arguments{
\item{id}{label id to update}

\item{label}{the label fields to update}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Get a specific label by id and user_id. \code{gm_update_label_patch()} is
identical to \code{gm_update_label()} but the latter uses \href{https://datatracker.ietf.org/doc/html/rfc5789}{HTTP PATCH} to allow partial
update.
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.labels/update}

\url{https://developers.google.com/gmail/api/reference/rest/v1/users.labels/patch}
}
\seealso{
Other label: 
\code{\link{gm_create_label}()},
\code{\link{gm_delete_label}()},
\code{\link{gm_labels}()},
\code{\link{gm_label}()}

Other label: 
\code{\link{gm_create_label}()},
\code{\link{gm_delete_label}()},
\code{\link{gm_labels}()},
\code{\link{gm_label}()}
}
\concept{label}
