% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{gmail_auth_config}
\alias{gmail_auth_config}
\title{View or set auth config}
\usage{
gmail_auth_config(app = NULL, secret_file = NULL)
}
\arguments{
\item{app}{OAuth app. Defaults to a tidyverse app that ships with gmailr.}

\item{secret_file}{Alternative way to bring your own app: the secret json
file downloaded from \url{https://console.cloud.google.com} containing
a client id and secret.}
}
\description{
This function gives advanced users more control over auth.
Whereas [gmail_auth()] gives control over tokens, `gmail_auth_config()`
gives control of:
  * The OAuth app. If you want to use your own app, setup a new project in
  [Google Developers Console](https://console.developers.google.com). Follow
  the instructions in
  [OAuth 2.0 for Mobile & Desktop Apps](https://developers.google.com/identity/protocols/OAuth2InstalledApp)
  to obtain your own client ID and secret. Provide these to
  [httr::oauth_app()].
}
\examples{
gmail_auth_config()
}
