% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmailr-package.R
\name{gmailr-configuration}
\alias{gmailr-configuration}
\alias{gm_default_email}
\title{Configuring gmailr}
\usage{
gm_default_email()
}
\description{
gmailr can be configured with various environment variables. Since gmailr
uses the gargle package to handle auth, gargle's configuration is also
relevant, which is mostly accomplished through \link[gargle:gargle_options]{options and associated accessor functions}.
}
\section{\code{gm_default_email()}}{

\code{gm_default_email()} returns the environment variable \code{GMAILR_EMAIL}, if it
exists, and \code{\link[gargle:gargle_options]{gargle::gargle_oauth_email()}}, otherwise.
}

\seealso{
Other auth functions: 
\code{\link{gm_auth_configure}()},
\code{\link{gm_auth}()},
\code{\link{gm_deauth}()},
\code{\link{gm_scopes}()}
}
\concept{auth functions}
