% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gmailr-deprecated-auth}
\alias{gmailr-deprecated-auth}
\alias{clear_token}
\alias{gmail_auth}
\alias{use_secret_file}
\title{Deprecated auth}
\usage{
clear_token()

gmail_auth(
  scope = c("read_only", "modify", "compose", "full"),
  id = the$id,
  secret = the$secret,
  secret_file = NULL
)

use_secret_file(filename)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are deprecated and no longer work.
}
\keyword{internal}
