% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gmailr-deprecated}
\alias{gmailr-deprecated}
\alias{id}
\title{These functions are deprecated and will be removed in a future release of
gmailr
#' \Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}}
\usage{
id(x, ...)
}
\description{
These functions are deprecated and will be removed in a future release of
gmailr
#' \Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
}
\keyword{internal}
