% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmailr-package.R
\docType{package}
\name{gmailr-package}
\alias{gmailr}
\alias{gmailr-package}
\title{gmailr: Access the 'Gmail' 'RESTful' API}
\description{
An interface to the 'Gmail' 'RESTful' API. Allows access to your 'Gmail' messages, threads, drafts and labels.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gmailr.r-lib.org}
  \item \url{https://github.com/r-lib/gmailr}
  \item Report bugs at \url{https://github.com/r-lib/gmailr/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co}

Authors:
\itemize{
  \item Jim Hester
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
