% Generated by roxygen2 (4.0.2): do not edit by hand
\name{history}
\alias{history}
\title{Retrieve change history for the inbox}
\usage{
history(start_history_id = NULL, num_results = NULL, label_id = NULL,
  page_token = NULL, user_id = "me")
}
\arguments{
\item{start_history_id}{the point to start the history.  The historyId can be obtained from a message, thread or previous list response.}

\item{num_results}{the number of results to return, max per page is 100}

\item{label_id}{filter history only for this label}

\item{page_token}{retrieve a specific page of results}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Retrieves the history results in chronological order
}
\examples{
\dontrun{
my_history = history("10")
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/history/list}
}

