% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmailr.R
\name{id}
\alias{id}
\title{Get the id of a gmailr object}
\usage{
id(x, ...)
}
\arguments{
\item{x}{the object from which to retrieve the id}

\item{...}{other parameters passed to methods}
}
\description{
Get the id of a gmailr object
}
\examples{
\dontrun{
id(my_message)
id(my_draft)
}
}

