% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{labels}
\alias{labels}
\title{Get a list of all labels}
\usage{
labels(user_id = "me")
}
\arguments{
\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Get a list of all labels for a user.
}
\examples{
\dontrun{
my_labels = labels()
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/labels/list}
}
\seealso{
Other label: \code{\link{create_label}},
  \code{\link{delete_label}}, \code{\link{label}},
  \code{\link{update_label}}
}

