% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{message}
\alias{message}
\title{Get a single message}
\usage{
message(id, user_id = "me", format = c("full", "metadata", "minimal",
  "raw"))
}
\arguments{
\item{id}{message id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}

\item{format}{format of the message returned}
}
\description{
Function to retrieve a given message by id
}
\examples{
\dontrun{
my_message = message(12345)
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/messages}
}
\seealso{
Other message: \code{\link{attachment}},
  \code{\link{delete_message}},
  \code{\link{import_message}},
  \code{\link{insert_message}}, \code{\link{messages}},
  \code{\link{modify_message}},
  \code{\link{save_attachments}},
  \code{\link{save_attachment}},
  \code{\link{send_message}}, \code{\link{trash_message}},
  \code{\link{untrash_message}}
}

