% Generated by roxygen2 (4.0.1): do not edit by hand
\name{messages_info}
\alias{messages_info}
\title{Extract information from a list of messages}
\usage{
messages_info(messages, what = "id")
}
\arguments{
\item{messages}{The output from \link{messages}.}

\item{what}{Either \code{'id'}, \code{'threadId'}, \code{'nextPageToken'},
\code{'resultSizeEstimate'}.}
}
\value{
For \code{what = 'id'} or \code{what = 'threadId'} a vector with the
message ids or thread ids, respectively. Otherwise the next page token or
result size estimate information.
}
\description{
Extract the message ids, thread ids, the next page token or the result size
estimate from a \code{gmail_messages} object created with \link{messages}.
}
\examples{
\dontrun{
#Search for R, return 10 results using label 1 including spam and trash folders
my_messages = messages("R", 10, "label_1", TRUE)

## Extract the ids from the messages
ids <- messages_info(my_messages)

## You can then use these ids as input for other functions.
}
}
\author{
L. Collado-Torres <lcollado@jhu.edu>
}
\seealso{
\link{messages}
}

