% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mime}
\alias{mime}
\title{Create a mime formatted message object}
\usage{
mime(..., body = NULL)
}
\arguments{
\item{...}{You can set header values initially, or use}

\item{body}{text to use for the body
\code{\link{common_fields}} set functions to set them after object creation.}
}
\description{
Create a mime formatted message object
}
\examples{
# using the field functions
msg = mime() \%>\%
 from("james.f.hester@gmail.com") \%>\%
 to("CRAN@R-project.org") \%>\%
 text_body("Please don't reject my package")

# alternatively you can set the fields using mime, however you hav()e
#  to use properly formatted MIME names
msg = mime(From="james.f.hester@gmail.com",
                   To="CRAN@R-project.org") \%>\%
        html_body("<b>Please<\\b> don't reject my package")
}
\seealso{
\code{\link{common_fields}}, \code{\link{body}}
}

