% Generated by roxygen2 (4.0.2): do not edit by hand
\name{text_body}
\alias{html_body}
\alias{text_body}
\title{Add a text or html body to a mime message}
\usage{
text_body(mime, body, ...)

html_body(mime, body, ...)
}
\arguments{
\item{mime}{message to add the body to}

\item{body}{the body to add to the message}

\item{...}{additional parameters to put in the attr field}
}
\description{
if called without a body returns the current body
}

