% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{mime_message}
\alias{mime_message}
\title{Create a mime formatted message object}
\usage{
mime_message(...)
}
\arguments{
\item{...}{You can set header values initially, or use
\code{\link{common_fields}} set functions to set them after object creation.}
}
\description{
Create a mime formatted message object
}
\examples{
# using the field functions
msg = mime_message() \%>\%
 from("james.f.hester@gmail.com") \%>\%
 to("CRAN@R-project.org") \%>\%
 text_body("Please don't reject my package")

# alternatively you can set the fields using mime_message, however you hav()e
#  to use properly formatted MIME names
msg = mime_message(From="james.f.hester@gmail.com",
                   To="CRAN@R-project.org") \%>\%
        html_body("<b>Please<\\b> don't reject my package")
}
\seealso{
\code{\link{common_fields}}, \code{\link{body}}
}

