% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{modify_thread}
\alias{modify_thread}
\title{Modify the labels on a thread}
\usage{
modify_thread(id, add_labels = character(0), remove_labels = character(0),
  user_id = "me")
}
\arguments{
\item{id}{thread id to access}

\item{add_labels}{labels to add to the specified thread}

\item{remove_labels}{labels to remove from the specified thread}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to modify the labels on a given thread by id.
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/threads/modify}
}

