% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%>\%}
\alias{\%>\%}
\title{Pipe statements}
\arguments{
\item{lhs,rhs}{A visualisation and a function to apply to it}
}
\description{
Like dplyr and ggvis gmailr also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
# Instead of
to(mime(), 'someone@somewhere.com')
# you can write
mime() \%>\% to('someone@somewhere.com')
}

