% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{print.gmail_draft}
\alias{print.gmail_draft}
\title{Print a gmail_draft}
\usage{
\method{print}{gmail_draft}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{other parameters passed to methods}
}
\description{
Print a gmail_draft
}
\examples{
\dontrun{
my_message
}
}

