% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{print.gmail_messages}
\alias{print.gmail_messages}
\title{Print a list of gmail_messages}
\usage{
\method{print}{gmail_messages}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{other parameters passed to methods}
}
\description{
Prints each message_id and the corresponding thread_id
}
\examples{
\dontrun{
my_message
}
}

