% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{print.gmail_threads}
\alias{print.gmail_threads}
\title{Print a list of gmail_threads}
\usage{
\method{print}{gmail_threads}(x, ...)
}
\arguments{
\item{x}{the object to print}

\item{...}{other parameters passed to methods}
}
\description{
Prints each thread_id and the corresponding snippet.
}
\examples{
\dontrun{
my_message
}
}

