% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/gmailr.R
\name{print}
\alias{print}
\alias{print.gmail_draft}
\alias{print.gmail_drafts}
\alias{print.gmail_message}
\alias{print.gmail_messages}
\alias{print.gmail_thread}
\alias{print.gmail_threads}
\title{Print gmailr objects}
\usage{
\method{print}{gmail_message}(x, ...)

\method{print}{gmail_draft}(x, ...)

\method{print}{gmail_drafts}(x, ...)

\method{print}{gmail_messages}(x, ...)

\method{print}{gmail_thread}(x, ...)

\method{print}{gmail_threads}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{additional arguments ignored}
}
\description{
Print gmailr objects
}

