% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quoted_printable_encode}
\alias{quoted_printable_encode}
\title{Encode text using quoted printable}
\usage{
quoted_printable_encode(data)
}
\arguments{
\item{data}{data to encode}
}
\description{
Does no do any line wrapping of the output to 76 characters
Implementation derived from the perl MIME::QuotedPrint
}
\references{
\url{http://search.cpan.org/~gaas/MIME-Base64-3.14/QuotedPrint.pm}
}

