% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save_attachment}
\alias{save_attachment}
\title{save the attachment to a file}
\usage{
save_attachment(x, filename)
}
\arguments{
\item{x}{attachment to save}

\item{filename}{location to save to}
}
\description{
this only works on attachments retrieved with \code{\link{attachment}}.
To save an attachment directly from a message see \code{\link{save_attachments}}
}
\examples{
\dontrun{
my_attachment = attachment('a32e324b', '12345')
save attachment to a file
save_attachment(my_attachment, 'photo.jpg')
}
}

