\name{scrape.cell}
\alias{scrape.cell}
\title{Scrape Cell Phone Extensions}
\usage{
  scrape.cell(URL = "http://www.emailtextmessages.com/")
}
\arguments{
  \item{URL}{The URL of the cell phone extension website to
  scrape from.}
}
\description{
  Grabs cell phone extensions from
  http://www.emailtextmessages.com/
}
\examples{
\dontrun{
scrape.cell()
}
}

