% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{send_draft}
\alias{send_draft}
\title{Send a draft}
\usage{
send_draft(draft, user_id = "me")
}
\arguments{
\item{draft}{the draft to send}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Send a draft to the recipients in the To, CC, and Bcc headers.
}
\examples{
\dontrun{
draft <- create_draft(mime(From="you@me.com", To="any@one.com",
                      Subject="hello", "how are you doing?"))
send_draft(draft)
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/drafts/send}
}
\seealso{
Other draft: \code{\link{drafts}}, \code{\link{draft}}
}

