% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{subject}
\alias{subject}
\alias{subject.gmail_draft}
\alias{subject.gmail_message}
\title{Get the subject field of a message or draft}
\usage{
subject(x, ...)

\method{subject}{gmail_message}(x, ...)

\method{subject}{gmail_draft}(x, ...)
}
\arguments{
\item{x}{the object from which to retrieve the field}

\item{...}{other parameters passed to methods}
}
\description{
Get the subject field of a message or draft
}
\examples{
\dontrun{
subject(my_message)
subject(my_draft)
}
}

