% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{thread}
\alias{thread}
\title{Get a single Thread.}
\usage{
thread(id, user_id = "me")
}
\arguments{
\item{id}{thread id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to retrieve a given thread by id
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/thread}
}

