% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{threads}
\alias{threads}
\title{Get a list of Threads.}
\usage{
threads(search = NULL, num_results = NULL, page_token = NULL,
  user_id = "me")
}
\arguments{
\item{search}{query to use, same format as gmail search box.}

\item{num_results}{the number of results to return, max per page is 100}

\item{page_token}{retrieve a specific page of results}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Get a list of threads possibly matching a given query string.
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/messages/list}
}

