% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{to}
\alias{to}
\alias{to.gmail_draft}
\alias{to.gmail_message}
\title{Get the to field of a message or draft}
\usage{
to(x, ...)

\method{to}{gmail_message}(x, ...)

\method{to}{gmail_draft}(x, ...)
}
\arguments{
\item{x}{the object from which to retrieve the field}

\item{...}{other parameters passed to methods}
}
\description{
Get the to field of a message or draft
}
\examples{
\dontrun{
to(my_message)
to(my_draft)
}
}

