% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{trash_message}
\alias{trash_message}
\title{Send a single message to the trash}
\usage{
trash_message(id, user_id = "me")
}
\arguments{
\item{id}{message id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to trash a given message by id.  This can be undone by \code{\link{untrash_message}}.
}
\examples{
\dontrun{
trash_message('12345')
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/messages/trash}
}
\seealso{
Other message: \code{\link{attachment}},
  \code{\link{delete_message}},
  \code{\link{import_message}},
  \code{\link{insert_message}}, \code{\link{messages}},
  \code{\link{message}}, \code{\link{modify_message}},
  \code{\link{save_attachments}},
  \code{\link{save_attachment}},
  \code{\link{send_message}}, \code{\link{untrash_message}}
}

