% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{trash_thread}
\alias{trash_thread}
\title{Send a single Thread to the trash.}
\usage{
trash_thread(id, user_id = "me")

trash_thread(id, user_id = "me")
}
\arguments{
\item{id}{thread id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to trash a given thread by id.  This can be undone by \code{\link{untrash_thread}}.

Function to untrash a given Thread by id.  This can reverse the results of a previous \code{\link{trash_thread}}.
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/threads/trash}

\url{https://developers.google.com/gmail/api/v1/reference/users/threads/untrash}
}

