% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/thread.R
\name{untrash_thread}
\alias{untrash_thread}
\title{Remove a single thread from the trash.}
\usage{
untrash_thread(id, user_id = "me")
}
\arguments{
\item{id}{thread id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to untrash a given thread by id.  This can reverse the results of a previous \code{\link{trash_thread}}.
}
\examples{
\dontrun{
untrash_thread(12345)
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/threads/untrash}
}

