% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/label.R
\name{update_label}
\alias{update_label}
\alias{update_label_patch}
\title{Update a existing label.}
\usage{
update_label(id, label, user_id = "me")

update_label_patch(id, label, user_id = "me")
}
\arguments{
\item{id}{label id to update}

\item{label}{the label fields to update}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Get a specific label by id and user_id.  \code{update_label_patch} is identical to \code{update_label} but the latter uses \href{http://tools.ietf.org/html/rfc5789}{HTTP PATCH} to allow partial update.
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/labels/update}

\url{https://developers.google.com/gmail/api/v1/reference/users/labels/patch}
}

