% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmailr.R
\name{use_secret_file}
\alias{use_secret_file}
\title{Use information from a secret file}
\usage{
use_secret_file(filename)
}
\arguments{
\item{filename}{the filename of the file}
}
\description{
This function sets the default secret and client_id to those in the secret
file
}

