% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmailr.R
\name{to}
\alias{to}
\alias{from}
\alias{cc}
\alias{bcc}
\alias{date}
\alias{subject}
\title{Methods to get values from message or drafts}
\usage{
to(x, ...)

from(x, ...)

cc(x, ...)

bcc(x, ...)

date(x, ...)

subject(x, ...)
}
\arguments{
\item{x}{the object from which to get or set the field}

\item{...}{other parameters passed to methods}
}
\description{
Methods to get values from message or drafts
}
