% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmailr.R
\name{to}
\alias{to}
\title{Methods to get values from message or drafts}
\usage{
to(x, ...)
}
\arguments{
\item{x}{the object from which to get or set the field}

\item{...}{other parameters passed to methods}
}
\description{
Methods to get values from message or drafts
}
\seealso{
\code{\link{common_fields}}
}

