% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to}
\alias{bcc.gmail_draft}
\alias{bcc.gmail_message}
\alias{cc.gmail_draft}
\alias{cc.gmail_message}
\alias{date}
\alias{date.gmail_draft}
\alias{date.gmail_message}
\alias{from}
\alias{from.gmail_message}
\alias{subject}
\alias{subject.gmail_draft}
\alias{subject.gmail_message}
\alias{to}
\alias{to.gmail_draft}
\alias{to.gmail_message}
\title{Methods to get values from message or drafts}
\usage{
to(x, ...)

\method{to}{gmail_message}(x, ...)

\method{to}{gmail_draft}(x, ...)

from(x, ...)

\method{from}{gmail_message}(x, ...)

cc.gmail_message(x, ...)

cc.gmail_draft(x, ...)

bcc.gmail_message(x, ...)

bcc.gmail_draft(x, ...)

date(x, ...)

\method{date}{gmail_message}(x, ...)

\method{date}{gmail_draft}(x, ...)

subject(x, ...)

\method{subject}{gmail_message}(x, ...)

\method{subject}{gmail_draft}(x, ...)
}
\arguments{
\item{x}{the object from which to get or set the field}

\item{...}{other parameters passed to methods}
}
\description{
Methods to get values from message or drafts
}
\seealso{
\code{\link{common_fields}}
}

