% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmailr.R
\name{gm_to}
\alias{gm_to}
\alias{gm_from}
\alias{gm_cc}
\alias{gm_bcc}
\alias{gm_date}
\alias{gm_subject}
\title{Methods to get values from message or drafts}
\usage{
gm_to(x, ...)

gm_from(x, ...)

gm_cc(x, ...)

gm_bcc(x, ...)

gm_date(x, ...)

gm_subject(x, ...)
}
\arguments{
\item{x}{the object from which to get or set the field}

\item{...}{other parameters passed to methods}
}
\description{
Methods to get values from message or drafts
}
