\name{anRpackage-package}
\alias{anRpackage-package}
\alias{anRpackage}
\docType{package}
\title{
What the package does (short line)
Send gmail with attachments
}
\description{
Send gmail with attachments
}
\details{
\tabular{ll}{
Package: \tab qdap\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-01\cr
License: \tab GPL-2\cr
}
}
\author{
Tyler Rinker

Maintainer: Tyler Rinker <tyler.rinker@gmail.com>
}

~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
