% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mime.R
\name{as.character.mime}
\alias{as.character.mime}
\title{Convert a mime object to character representation}
\usage{
\method{as.character}{mime}(x, newline = "\\r\\n", ...)
}
\arguments{
\item{x}{object to convert}

\item{newline}{value to use as newline character}

\item{...}{further arguments ignored}
}
\description{
This function converts a mime object into a character vector
}
