% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/mime.R
\name{attach_part}
\alias{attach_file}
\alias{attach_part}
\title{Attach an object to a mime message}
\usage{
attach_part(mime, body, ...)

attach_file(mime, filename, type = NULL, ...)
}
\arguments{
\item{mime}{object to attach to}

\item{body}{data to attach}

\item{...}{additional arguments put into the attr field of the object}

\item{filename}{name of file to attach}

\item{type}{mime type of the attached file}
}
\description{
Attach an object to a mime message
}

