% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmailr.R
\name{body}
\alias{body}
\title{Get the body text of a message or draft}
\usage{
body(x, ...)
}
\arguments{
\item{x}{the object from which to retrieve the body}

\item{...}{other parameters passed to methods}
}
\description{
Get the body text of a message or draft
}
\examples{
\dontrun{
body(my_message)
body(my_draft)
}
}

