% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/mime.R
\name{to.mime}
\alias{bcc}
\alias{cc}
\alias{common_fields}
\alias{from.mime}
\alias{subject.mime}
\alias{to.mime}
\title{Accessor functions for mime messages}
\usage{
\method{to}{mime}(x, vals, ...)

\method{from}{mime}(x, val, ...)

cc(x, vals, ...)

bcc(x, vals, ...)

\method{subject}{mime}(x, val, ...)
}
\arguments{
\item{x}{the object whose fields you are setting}

\item{vals}{one or more values to use, will be joined by commas}

\item{...}{other arguments ignored}

\item{val}{the value to set}
}
\description{
Accessor functions for mime messages
}

