% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete_label}
\alias{delete_label}
\title{Permanently delete a label}
\usage{
delete_label(id, user_id = "me")
}
\arguments{
\item{id}{label id to retrieve}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to delete a label by id.  This cannot be undone!
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/labels/delete}
}

