% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Get a single draft}
\usage{
draft(id, user_id = "me", format = c("full", "minimal", "raw"))
}
\arguments{
\item{id}{draft id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}

\item{format}{format of the draft returned}
}
\description{
Function to retrieve a given draft by id
}
\examples{
\dontrun{
my_draft = draft('12345')
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/drafts/get}
}

