% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{drafts}
\alias{drafts}
\title{Get a list of drafts}
\usage{
drafts(num_results = NULL, page_token = NULL, user_id = "me")
}
\arguments{
\item{num_results}{the number of results to return.}

\item{page_token}{retrieve a specific page of results}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Get a list of drafts possibly matching a given query string.
}
\examples{
\dontrun{
my_drafts = drafts()

first_10_drafts = drafts(10)
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/drafts/list}
}
\seealso{
Other draft: \code{\link{draft}}, \code{\link{send_draft}}
}

