% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_label.R
\name{gm_labels}
\alias{gm_labels}
\title{Get a list of all labels}
\usage{
gm_labels(user_id = "me")
}
\arguments{
\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Get a list of all labels for a user.
}
\examples{
\dontrun{
my_labels <- gm_labels()
}
}
\references{
\url{https://developers.google.com/gmail/api/reference/rest/v1/users.labels/list}
}
\seealso{
Other label: 
\code{\link{gm_create_label}()},
\code{\link{gm_delete_label}()},
\code{\link{gm_label}()},
\code{\link{gm_update_label}()}
}
\concept{label}
