% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_auth.R
\name{gm_token}
\alias{gm_token}
\title{Produce configured token}
\usage{
gm_token()
}
\value{
A \code{request} object (an S3 class provided by \link[httr:httr-package]{httr}).
}
\description{
For internal use or for those programming around the Gmail API.
Returns a token pre-processed with \code{\link[httr:config]{httr::config()}}. Most users
do not need to handle tokens "by hand" or, even if they need some
control, \code{\link[=gm_auth]{gm_auth()}} is what they need. If there is no current
token, \code{\link[=gm_auth]{gm_auth()}} is called to either load from cache or
initiate OAuth2.0 flow.
If auth has been deactivated via \code{\link[=gm_deauth]{gm_deauth()}}, \code{gm_token()}
returns \code{NULL}.
}
\examples{
\dontshow{if (gm_has_token()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
gm_token()
\dontshow{\}) # examplesIf}
}
\seealso{
Other low-level API functions: 
\code{\link{gm_has_token}()}
}
\concept{low-level API functions}
