% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmailr.R
\name{gmail_auth}
\alias{gmail_auth}
\title{Setup oauth authentication for your gmail}
\usage{
gmail_auth(id = gmailr_env$id, secret = gmailr_env$secret,
  scope = c("read_only", "modify", "compose", "full"))
}
\arguments{
\item{scope}{the authentication scope to use}

\item{secret_file}{the secret json file downloaded from \url{https://cloud.google.com/console#/project}}
}
\description{
Setup oauth authentication for your gmail
}
\examples{
\dontrun{
gmail_auth("file", "compose")
}
}

