% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-auth.R
\name{gmail_auth}
\alias{gmail_auth}
\title{Setup oauth authentication for your gmail}
\usage{
gmail_auth(
  scope = c("read_only", "modify", "compose", "full"),
  id = the$id,
  secret = the$secret,
  secret_file = NULL
)
}
\arguments{
\item{scope}{the authentication scope to use}

\item{id}{the client_id to use for authentication}

\item{secret}{the client secret to use for authentication}

\item{secret_file}{the secret json file downloaded from \url{https://console.cloud.google.com}}
}
\description{
Setup oauth authentication for your gmail
}
\examples{
\dontrun{
gmail_auth("compose")
}
}
\seealso{
use_secret_file to set the default id and secret to a different
value than the default.
}
\keyword{internal}
