% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gmailr-deprecated}
\alias{gmailr-deprecated}
\alias{id}
\alias{to}
\alias{from}
\alias{cc}
\alias{bcc}
\alias{date}
\alias{subject}
\alias{body}
\alias{message}
\alias{messages}
\alias{trash_message}
\alias{untrash_message}
\alias{delete_message}
\alias{modify_message}
\alias{attachment}
\alias{save_attachment}
\alias{save_attachments}
\alias{insert_message}
\alias{import_message}
\alias{send_message}
\alias{threads}
\alias{thread}
\alias{trash_thread}
\alias{untrash_thread}
\alias{delete_thread}
\alias{modify_thread}
\alias{draft}
\alias{drafts}
\alias{create_draft}
\alias{send_draft}
\alias{labels}
\alias{label}
\alias{update_label}
\alias{update_label_patch}
\alias{delete_label}
\alias{create_label}
\alias{history}
\alias{mime}
\alias{text_body}
\alias{html_body}
\alias{attach_part}
\alias{attach_file}
\title{These functions are deprecated and will be removed in a future release of
gmailr
#' \Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}}
\usage{
id(x, ...)

to(x, ...)

from(x, ...)

cc(x, ...)

bcc(x, ...)

date(x, ...)

subject(x, ...)

body(x, ...)

message(...)

messages(...)

trash_message(...)

untrash_message(...)

delete_message(...)

modify_message(...)

attachment(...)

save_attachment(...)

save_attachments(...)

insert_message(...)

import_message(...)

send_message(...)

threads(...)

thread(...)

trash_thread(...)

untrash_thread(...)

delete_thread(...)

modify_thread(...)

draft(...)

drafts(...)

create_draft(...)

send_draft(...)

labels(...)

label(...)

update_label(...)

update_label_patch(...)

delete_label(...)

create_label(...)

history(...)

mime(...)

text_body(...)

html_body(...)

attach_part(...)

attach_file(...)
}
\description{
These functions are deprecated and will be removed in a future release of
gmailr
#' \Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
}
\keyword{internal}
