% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gmailr-deprecated}
\alias{gmailr-deprecated}
\alias{id}
\alias{to}
\alias{from}
\alias{cc}
\alias{bcc}
\alias{date}
\alias{subject}
\alias{body}
\alias{message}
\alias{messages}
\alias{trash_message}
\alias{untrash_message}
\alias{delete_message}
\alias{modify_message}
\alias{attachment}
\alias{save_attachment}
\alias{save_attachments}
\alias{insert_message}
\alias{import_message}
\alias{send_message}
\alias{threads}
\alias{thread}
\alias{trash_thread}
\alias{untrash_thread}
\alias{delete_thread}
\alias{modify_thread}
\alias{draft}
\alias{drafts}
\alias{create_draft}
\alias{send_draft}
\alias{labels}
\alias{label}
\alias{update_label}
\alias{update_label_patch}
\alias{delete_label}
\alias{create_label}
\alias{history}
\alias{mime}
\alias{text_body}
\alias{html_body}
\alias{attach_part}
\alias{attach_file}
\title{Deprecated functions}
\usage{
id(x, ...)

to(x, ...)

from(x, ...)

cc(x, ...)

bcc(x, ...)

date(x, ...)

subject(x, ...)

body(x, ...)

message(...)

messages(...)

trash_message(...)

untrash_message(...)

delete_message(...)

modify_message(...)

attachment(...)

save_attachment(...)

save_attachments(...)

insert_message(...)

import_message(...)

send_message(...)

threads(...)

thread(...)

trash_thread(...)

untrash_thread(...)

delete_thread(...)

modify_thread(...)

draft(...)

drafts(...)

create_draft(...)

send_draft(...)

labels(...)

label(...)

update_label(...)

update_label_patch(...)

delete_label(...)

create_label(...)

history(...)

mime(...)

text_body(...)

html_body(...)

attach_part(...)

attach_file(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All of these functions have been renamed, by adding a \code{gm_} prefix. This is
to eliminate name conflicts with other packages (including the base package).
Starting with gmailr 1.0.0 (released 2019-08-23), the deprecated functions
generated a warning. As of gmailr 2.0.0, they generate an error and they will
be completely removed in a future release.
}
\keyword{internal}
