% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/gmailr.R
\docType{package}
\name{gmailr}
\alias{gmailr}
\alias{gmailr-package}
\title{\pkg{gmailr} makes gmail access easy.}
\description{
\code{gmailr} provides an interface to the gmail api \url{https://developers.google.com/gmail/api/}
}

