% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{id}
\alias{id}
\title{Get the id of a message or draft}
\usage{
id(x, ...)
}
\arguments{
\item{x}{the object from which to retrieve the id}

\item{...}{other parameters passed to methods}
}
\description{
Get the id of a message or draft
}
\examples{
\dontrun{
id(my_message)
id(my_draft)
}
}

