% Generated by roxygen2 (4.0.2): do not edit by hand
\name{label}
\alias{label}
\title{Get a specific label}
\usage{
label(id, user_id = "me")
}
\arguments{
\item{id}{label id to retrieve}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Get a specific label by id and user_id.
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/labels/get}
}

