% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{labels}
\alias{labels}
\title{Get a list of all labels}
\usage{
labels(user_id = "me")
}
\arguments{
\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Get a list of all labels for a user.
}
\examples{
\dontrun{
my_labels = labels()
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/labels/list}
}

