% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/message.R
\name{message}
\alias{message}
\title{Get a single message}
\usage{
message(id, user_id = "me", format = c("full", "minimal", "raw"))
}
\arguments{
\item{id}{message id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}

\item{format}{format of the message returned}
}
\description{
Function to retrieve a given message by id
}
\examples{
\dontrun{
my_message = message(12345)
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/messages}
}

