% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/message.R
\name{modify_message}
\alias{modify_message}
\title{Modify the labels on a message}
\usage{
modify_message(id, add_labels = character(0), remove_labels = character(0),
  user_id = "me")
}
\arguments{
\item{id}{message id to access}

\item{add_labels}{label IDs to add to the specified message}

\item{remove_labels}{label IDs to remove from the specified message}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to modify the labels on a given message by id.  Note you need to
use the label ID as arguments to this function, not the label name.
}
\examples{
\dontrun{
modify_message(12345, add_labels='label_1')
modify_message(12345, remove_labels='label_1')
#add and remove at the same time
modify_message(12345, add_labels='label_2', remove_labels='label_1')
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/messages/modify}
}

