% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmailr.R, R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe statements}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs, rhs}{A visualisation and a function to apply to it}
}
\description{
Like dplyr and ggvis gmailr also uses the pipe function, \verb{\\\%>\\\%} to turn
function composition into a series of imperative statements.
}
\examples{
# Instead of
gm_to(gm_mime(), 'someone@somewhere.com')
# you can write
gm_mime() \%>\% gm_to('someone@somewhere.com')
}
\keyword{internal}
