% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/draft.R
\name{send_draft}
\alias{send_draft}
\title{Send a draft}
\usage{
send_draft(id, upload_type = c("media", "multipart", "resumable"),
  user_id = "me")
}
\arguments{
\item{id}{the draft id to send}

\item{upload_type}{type of upload request}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Send a draft to the recipients in the To, CC, and Bcc headers.
}
\examples{
\dontrun{
send_draft(12345)
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/drafts/send}
}

