% Generated by roxygen2 (4.0.2): do not edit by hand
\name{send_message}
\alias{send_message}
\title{Send a message from a mime message}
\usage{
send_message(mail, user_id = "me", label_ids = NULL, type = c("multipart",
  "media", "resumable"))
}
\arguments{
\item{mail}{mime mail message created by mime}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}

\item{label_ids}{optional label ids to apply to the message}

\item{type}{the type of upload to perform}
}
\description{
Send a message from a mime message
}
\examples{
\dontrun{
send_message(mime(from="you@me.com", to="any@one.com",
                          subject='hello", "how are you doing?"))
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/messages/send}
}

