% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{thread}
\alias{thread}
\title{Get a single thread}
\usage{
thread(id, user_id = "me")
}
\arguments{
\item{id}{thread id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to retrieve a given thread by id
}
\examples{
\dontrun{
my_thread = thread(12345)
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/threads}
}
\seealso{
Other thread: \code{\link{delete_thread}},
  \code{\link{modify_thread}}, \code{\link{threads}},
  \code{\link{trash_thread}}, \code{\link{untrash_thread}}
}

