% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/message.R
\name{trash_message}
\alias{trash_message}
\title{Send a single message to the trash}
\usage{
trash_message(id, user_id = "me")
}
\arguments{
\item{id}{message id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to trash a given message by id.  This can be undone by \code{\link{untrash_message}}.
}
\examples{
\dontrun{
trash_message('12345')
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/messages/trash}
}

