% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{trash_thread}
\alias{trash_thread}
\title{Send a single thread to the trash}
\usage{
trash_thread(id, user_id = "me")
}
\arguments{
\item{id}{thread id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to trash a given thread by id.  This can be undone by \code{\link{untrash_thread}}.
}
\examples{
\dontrun{
trash_thread(12345)
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/threads/trash}
}
\seealso{
Other thread: \code{\link{delete_thread}},
  \code{\link{modify_thread}}, \code{\link{threads}},
  \code{\link{thread}}, \code{\link{untrash_thread}}
}

