% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{untrash_message}
\alias{untrash_message}
\title{Remove a single message from the trash}
\usage{
untrash_message(id, user_id = "me")
}
\arguments{
\item{id}{message id to access}

\item{user_id}{gmail user_id to access, special value of 'me' indicates the authenticated user.}
}
\description{
Function to trash a given message by id.  This can be undone by \code{\link{untrash_message}}.
}
\examples{
\dontrun{
untrash_message('12345')
}
}
\references{
\url{https://developers.google.com/gmail/api/v1/reference/users/messages/trash}
}

